+++ app/views/tpi/companies/show.html.erb
@@ -1,5 +1,5 @@
 <h1>Company Details</h1>
-<%= link_to 'back to all Sectors', sectors_path %>
+<%= link_to 'back to all Sectors', tpi_sectors_path %>
 
 <hr>
 
@@ -64,12 +64,11 @@
 
 <hr>
 
-<!-- COMPANIES -->
 <h3>Carbon Performance <%= @company.name %></h3>
 <p>Carbon Performance alignment of <%= @company.name %> with the Paris agreement benchmarks.</p>
 <div class="chart">
 <%= line_chart(
-      emissions_chart_data_company_path(@company.id),
+      emissions_chart_data_tpi_company_path(@company.id),
       width: '800px',
       height: '600px',
       library: default_chart_options
