+++ config/routes.rb
@@ -1,18 +1,19 @@
 Rails.application.routes.draw do
-  # TPI
-  resources :sectors, only: [:show, :index] do
-    collection do
-      get :levels_chart_data
-      get :benchmarks_chart_data
+  namespace :tpi do
+    resources :sectors, only: [:show, :index] do
+      collection do
+        get :levels_chart_data
+        get :benchmarks_chart_data
+      end
+      member do
+        get :levels_chart_data
+        get :emissions_chart_data
+      end
     end
-    member do
-      get :levels_chart_data
-      get :emissions_chart_data
-    end
-  end
-  resources :companies, only: [:show] do
-    member do
-      get :emissions_chart_data
+    resources :companies, only: [:show] do
+      member do
+        get :emissions_chart_data
+      end
     end
   end
